Moved the Isotope structs into their own files to make it easier to get around
Added `use` to work with disposable resources and clean-up properly.

Added `use` to work with a dispose function to and clean-up properly.

Added withWebDriver that allows a computation to be used in context

Added withChromeDriver, withEdgeDriver, withFirefoxDriver

Removed the Run variants that take a IWebDriver
